init `configuration` file for layout
